pyproject.toml
[tool.poetry]
name = "food-formulations-pfi"
version = "0.1.0"
description = "MLOps platform for vegan food formulations"
authors = ["Martin Berns <tu@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"  # Cambiamos ^3.0.0 por ^2.2.0
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
sqlalchemy = "^2.0.46"
mlflow = "^3.8.1"
scikit-learn = "^1.5.0"
xgboost = "^3.1.3"
lightgbm = "^4.0.0"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
streamlit = "^1.53.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
